multi-listener

+++ physic.h
@@ -5,7 +5,7 @@
 #include <QString>
 
 // Definition de constantes
-const int VITESSE_SON = 340; // m/s
+const float VITESSE_SON = 340; // m/s
 
 // Classes
 class Material
